+++ src/transpiler/tests/csout/test45a.cs
@@ -29,12 +29,12 @@ public static partial class Program
 {
     public static void LoadFlags(string cFile, ref dynamic[] aFlag, decimal nCount = default)
     {
-        decimal i = default;
+        int i = default;
 
         HbRuntime.ASize(ref aFlag, nCount);
         for (i = 1; i <= nCount; i++)
         {
-            aFlag[(int)(i) - 1] = cFile + ":" + HbRuntime.Str(i, 1);
+            aFlag[i - 1] = cFile + ":" + HbRuntime.Str(i, 1);
         }
 
         return;
