+++ src/transpiler/tests/csout/test46a.cs
@@ -16,12 +16,12 @@ public static partial class Program
 {
     public static void SaveFlags(string cFile, ref dynamic[] aFlag, decimal nCount = default)
     {
-        decimal i = default;
+        int i = default;
 
         HbRuntime.ASize(ref aFlag, nCount);
         for (i = 1; i <= nCount; i++)
         {
-            aFlag[(int)(i) - 1] = cFile + "#" + HbRuntime.Str(i, 1);
+            aFlag[i - 1] = cFile + "#" + HbRuntime.Str(i, 1);
         }
 
         return;
